requirements.txt
# Instructions for creating a Python 3.9 environment and installing necessary packages
# 1. Create the virtual environment:
#    conda create -n <name> python=3.9 -y
#    (Replace <name> with your desired environment name)
#
# 2. Activate the virtual environment:
#    conda activate <name>
#
# 3. Install PyTorch and CUDA Toolkit 11.7 (pytorch=2.0.1=py3.9_cuda11.7):
#    conda install pytorch cudatoolkit=11.7 -c pytorch

#
# Afterward, install the required packages using the following command:
#    pip install -U -r requirements.txt


pytorch-lightning==2.0.4
autoreject==0.4.3
einops==0.7.0
huggingface-hub==0.21.4
librosa==0.10.1
matplotlib==3.8.4
mne==1.6.1
omegaconf==2.3.0
openpyxl==3.1.2
transformers==4.38.2
scheduling-utils==0.2.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.0
